refactor(aws): split ec2 runner module
