+++ lambdas/functions/control-plane/src/aws/ec2-runners.d.ts
@@ -7,7 +7,7 @@ import {
   Placement,
   FleetBlockDeviceMappingRequest,
 } from '@aws-sdk/client-ec2';
-import { LambdaRunnerSource } from '../scale-runners/scale-up';
+import type { LambdaRunnerSource } from '../scale-runners/types';
 
 export type RunnerType = 'Org' | 'Repo';
