+++ lambdas/functions/control-plane/src/aws/ec2-runners.test.ts
@@ -22,9 +22,9 @@ import 'aws-sdk-client-mock-jest/vitest';
 
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import ScaleError from './../scale-runners/ScaleError';
-import { createRunner, listEC2Runners, tag, terminateRunner, untag } from './runners';
-import type { Ec2OverrideConfig, RunnerInfo, RunnerInputParameters, RunnerType } from './runners.d';
-import { LambdaRunnerSource } from '../scale-runners/scale-up';
+import { createRunner, listEC2Runners, tag, terminateRunner, untag } from './ec2-runners';
+import type { Ec2OverrideConfig, RunnerInfo, RunnerInputParameters, RunnerType } from './ec2-runners.d';
+import type { LambdaRunnerSource } from '../scale-runners/types';
 
 process.env.AWS_REGION = 'eu-east-1';
 const mockEC2Client = mockClient(EC2Client);
